First Commit
First version of working procedural terrain using Unity's built in
Mathf.PerlinNoise() function. Has three separate terrain calculations:
p1 (ground), p2 (rocks), and p3 (hills). Other terrain aspects such as
mountains or valleys can be added easily. The current map is made of 9
terrains with 3 on each side. The dimension of the whole map is 500x500
units, with a max height of 1000 units. To change these values make sure
to apply ratio of change on the height to the scale of the noise, and to
apply the ratio of change on the dimension to the frequency of noise.